Code/requirements.txt
streamlit
google-generativeai
python-dotenv
tiktoken
openai
wheel
langchain-text-splitters
langchain
langchain-community
langchain-openai
